chore: update example environment file
The example environment file (.example.env) was updated to include a comment with instructions on how to fill it out and update the secrets. This change was made to improve the usability of the file.

feat: add app configuration file

A new file (app.config.ts) was added to the project. This file contains the configuration for the Expo app, including the name, version, orientation, icons, splash screen, asset bundle patterns, iOS and Android specific configurations, web bundler settings, plugins, experiments, and extra settings. This change was made to provide a centralized location for app configuration and improve maintainability.

refactor: remove obsolete app.json file

The app.json file was removed from the project as it is no longer needed. The configuration settings from app.json were migrated to the new app.config.ts file. This change was made to simplify the project structure and remove redundant configuration files.

chore: update Atlas configuration

The appId in the atlasConfig.json file was updated to "breathe-easy-sync-db-bbiwm". This change was made to reflect the correct appId for the Atlas configuration.

chore: update package name in package.json

The name field in the package.json file was updated to "breathe-easy". This change was made to reflect the correct package name for the project.