feat(components): add new Styled component and export it in index.ts
The `Styled.tsx` file was added, which contains two new components: `StyledButton` and `StyledLinkButton`. These components are used to create styled buttons and link buttons with customizable variants and color tiers. The `StyledButton` component is a button with customizable text and background color, while the `StyledLinkButton` component is a link button that wraps the `StyledButton` component with a link from the `expo-router` package.

The `StyledText.tsx` file was deleted, as it is no longer needed.

In the `Themed.tsx` file, the `Button` component was added. This component is used to create customizable buttons with different variants (contained, outlined, text) and color tiers. The `Button` component uses the `Pressable` component from `react-native` to handle button press events.

In the `index.ts` file, the `Styled` component was added to the list of exported components.