feat(hooks): add useUserData hook to handle user data operations
The useUserData hook is added to handle user data operations. It provides functions to read, initialize, and edit user data. The hook is used in the useNotifications hook to handle the expoPushToken. The useCustomUserData hook is removed as it is no longer needed.

BREAKING CHANGE: The useCustomUserData hook is removed. Use the useUserData hook instead.